Fixed money issue in schema

+++ db/schema/02_listings.sql
@@ -4,7 +4,7 @@ CREATE TABLE listings (
   user_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
   title VARCHAR(255) NOT NULL,
   description TEXT NOT NULL,
-  price INTEGER,
+  price MONEY,
   created_at TIMESTAMP,
   sold_at TIMESTAMP
 );
